Add files via upload
This was the finished prototype (without the driver chip for the coil igniter). 
The LM358 op-amps were used to condition the signals from the cam and crank sensors (both inductive sensors). Not only to convert the balanced input to an unbalanced output, but also amplifying such that a suitable trigger level would be output to the arduino (the Schmidt trigger part). The pots were used to adjust (fine-tune) the voltage. NPN transistors were used to isolate the arduino output from the ECU (where the N signal was input to). 
In this model, LEDs were also added to indicate the status of some of the lines (eg. IGT, IGF, N, etc.).