Add files via upload
Basic idea here. 
The schmidt trigger was implemented through slewing an lm358 op-amp in an inverting amplifier configuration with an offset (level) adjustment; then through a buffer-amplifier. 10-turn precision pots were used to adjust the offset voltage (input to + of op-amp to perform inverting amplification) and to adjust the amplification feedback (between - and Vout of same op-amp). These pots were used for purposes of convenience (fine adjustment possible and easy to do). For deployment, measure the resistance of the pot, and buy the appropriate value (to place into a PCB). 

The schematic shown is a basic idea and is (then as now) eligible for many upgrades and improvements. It did work and was easy to adjust.